Add contrasting rain effect for bright background
The default rain was appearing great on the dark
background, but not when it was daytime, and the
background was yellow.

Depending on the time of day (and if it's raining)
the contrast will be different.

Also tidied up the CSS a little, moving the body
properties needed for the rain effect to where the
body is declared in App.css